Restructure entire git repo
Seperated the original callon file into two files `callon` and
`callon.js`. The `callon` file pulls in student data, instantiates the
CallOn object defined in the `callon.js` file.

The `callon.js` file now contains the bulk of the previous
functionality. A `CallOn` object is defined providing a building point
for future funtionality.

The `test.js` file is far from comprehensive. It was used to make sure
that the `.groupsOf` was working correctly.

The keys of the `student.json` file are the students names. The values
are meaningless as of now, but the idea is to store statistics related
the students in regards to the script. (things like how many times the
student had been called and what not)